implemented basic type of word signature
For each word, I average the fourier spectrum. I then output the difference between the current average and the last average. So far, the program seems pretty good at differentiating the current word from the last word; it can tell (with a decent margin of certainty) when a word is repeated vs. when a different word is said.