updated training and audio parameters
Now narrower time windows are used for audio sequences. This theoretically allows the network to be more precise while classifying audio, although it is unclear whether doing so actually helps. I mainly chose 10ms and 5ms because those are the times the first CTC paper used.

The network is also more powerful now, with a hidden layer for processing the output. I did this because, without a second output layer, certain bidirectional functions like bidirectional XOR are not possible. The second layer allows the network to learn more possible functions of the forward/backward layers.

Another new addition is that regularization is done by adding gaussian noise to the input signal. This is done to mimick the original CTC paper, which did something similar.